+++ Sources/BaseFoundationiOS/Templates/Types/AsyncHandler.swift
@@ -1,0 +1,10 @@
+//
+//  AsyncHandler.swift
+//  
+//
+//  Created by pablo borquez on 21/07/23.
+//
+
+import Foundation
+
+public typealias AsyncHandler = @Sendable () async -> Void